backend/requirements.txt
# Lint
flake8

# Testing dependencies
pytest
pytest-cov

# Dependencies
Django

#database
sqlalchemy
mysqlclient

# Image reading
pytesseract
Pillow

# Data parsing
pandas
